# public/swagger-ui/swagger.yaml
openapi: 3.0.0
info:
  title: 'Manavola Api'
  version: '0.1'
servers:
  -
    url: 'http://manavolabackend-env.eba-sweq2qp9.us-east-2.elasticbeanstalk.com'
    description: 'Manavola backend'
paths:
...
  /register:
    post:
      tags:
        - Register
      description: 'register a new user'
      operationId: 023bf4410e083690ec4c5e3227c07e2d
      requestBody:
        request: 'create new user'
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
                - username
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                username:
                  type: string
                parrainerId:
                  type: string
              type: object
      responses:
        '200':
          description: 'data received when in success'
          content:
            application/json:
              schema:
                properties:
                  user_id: { type: integer }
                  email: { type: string, format: email }
                  username: { type: string }
                  parrainageId: { type: string }
                type: object
  /registerverification:
    post:
      tags:
        - Register
      description: 'to make the user verified'
      operationId: acad9a7252c5a1e55f1a6c295e62a391
      requestBody:
        request: 'register verification'
        description: 'data to enter for making the user verified'
        required: true
        content:
          multipart/form-data:
            schema:
              required:
                - rectophoto
                - versoPhoto
                - selfieAvecCIN
              properties:
                rectophoto:
                  type: file
                  format: binary
                versoPhoto:
                  type: file
                  format: binary
                selfieAvecCIN:
                  type: file
                  format: binary
              type: object
          application/json:
            schema:
              required:
                - numeroCIN
                - fullname
              properties:
                numeroCIN:
                  type: string
                email:
                  type: string
                  format: email
                fullname:
                  type: string
              type: object
      responses:
        '201':
          description: Created
...